# backend/openapi.yaml
openapi: 3.0.0
info:
  version: 2.0.1
  title: Anki Clone Application API
  description: This is the API of the Anki Clone Application. It allows you to manage users, learning packages, learning facts, tags and statistics.
paths:
...
  /api/user/{user_id}:
    get:
      summary: Get a user by id
      operationId: getUserById
      tags:
        - Users
      parameters:
        - name: user_id
          in: path
          description: The id of the user to retrieve
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    put:
      summary: Update a user by id
      operationId: updateUserById
      tags:
        - Users
      parameters:
        - name: user_id
          in: path
          description: The id of the user to update
          required: true
          schema:
            type: number
      requestBody:
        description: The user to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                  description: The mail of the user
                  example: john.doe@gmail.com
                pseudo:
                  type: string
                  description: The pseudo of the user
                  example: John
                password:
                  type: string
                  description: The password of the user
                  example: 123456
                profilePicture:
                  type: string
                  description: The profile picture of the user
                  example: https://www.google.com/
      responses:
        '204':
          description: Updated
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
  /api/package:
    get:
      summary: Get all learning packages
      operationId: getLearningPackage
      tags:
        - Learning Packages
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LearningPackage'
        '500':
          description: Internal Server Error
    post:
      summary: Create a new learning package
      operationId: createLearningPackage
      tags:
        - Learning Packages
      requestBody:
        description: The learning package to create
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: The title of the learning package
                  example: Learn TypeScript
                  required: true
                description:
                  type: string
                  description: The description of the learning package
                  example: A comprehensive guide to understanding TypeScript, a statically typed superset of JavaScript.
                  required: false
                category:
                  type: string
                  description: The category of the learning package
                  example: Programming Language
                  required: false
                targetAudience:
                  type: string
                  description: The target audience of the learning package
                  example: Developers familiar with JavaScript
                  required: false
                duration:
                  type: number
                  description: The duration of the learning package in hours
                  example: 2
                  required: false
                creatorId:
                  type: number
                  description: The id of the creator of the learning package
                  example: 1
                  required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningPackage'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /api/package/summary:
    get:
      summary: Get all learning packages summary
      operationId: getLearningPackageSummary
      tags:
        - Learning Packages
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    packageId:
                      type: number
                      description: The id of the learning package
                      example: 1
                    title:
                      type: string
                      description: The title of the learning package
                      example: Learn TypeScript
        '500':
          description: Internal Server Error
  /api/package/search:
    get:
      summary: Search learning packages by title or tag
      operationId: searchLearningPackage
      tags:
        - Learning Packages
      parameters:
        - name: title
          in: query
          description: The title of the learning package
          required: false
          schema:
            type: string
        - name: tag
          in: query
          description: The tag of the learning package
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LearningPackage'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /api/package/search-title:
    post:
      summary: Search learning packages by title
      operationId: searchLearningPackageByTitle
      tags:
        - Learning Packages
      requestBody:
        description: The title of the learning package
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: The title of the learning package
                  example: Introduction to HTML
                  required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LearningPackage'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /api/package/{package_id}:
    get:
      summary: Get a learning package by id
      operationId: getLearningPackageById
      tags:
        - Learning Packages
      parameters:
        - name: package_id
          in: path
          description: The id of the learning package
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningPackage'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    put:
      summary: Update a learning package by id
      operationId: updateLearningPackageById
      tags:
        - Learning Packages
      parameters:
        - name: package_id
          in: path
          description: The id of the learning package
          required: true
          schema:
            type: number
      requestBody:
        description: The learning package to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: The title of the learning package
                  example: Learn TypeScript
                description:
                  type: string
                  description: The description of the learning package
                  example: A comprehensive guide to understanding TypeScript, a statically typed superset of JavaScript.
                category:
                  type: string
                  description: The category of the learning package
                  example: Programming Language
                targetAudience:
                  type: string
                  description: The target audience of the learning package
                  example: Developers familiar with JavaScript
                duration:
                  type: number
                  description: The duration of the learning package in hours
                  example: 2
      responses:
        '204':
          description: Updated
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    delete:
      summary: Delete a learning package by id
      operationId: deleteLearningPackageById
      tags:
        - Learning Packages
      parameters:
        - name: package_id
          in: path
          description: The id of the learning package
          required: true
          schema:
            type: number
      responses:
        '204':
          description: Deleted
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

  # LEARNING FACT API
...
  /api/package/{package_id}/fact:
    get:
      summary: Get all learning facts by learning package id
      operationId: getLearningFactByLearningPackageId
      tags:
        - Learning Packages <> Learning Facts
      parameters:
        - name: package_id
          in: path
          description: The id of the learning package
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LearningFact'
        '500':
          description: Internal Server Error
    post:
      summary: Create a new learning fact
      operationId: createLearningFact
      tags:
        - Learning Packages <> Learning Facts
      parameters:
        - name: package_id
          in: path
          description: The id of the learning package
          required: true
          schema:
            type: number
      requestBody:
        description: The learning fact to create
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                front:
                  type: string
                  description: The front of the learning fact
                  example: What is TypeScript?
                  required: true
                back:
                  type: string
                  description: The back of the learning fact
                  example: TypeScript is a statically typed superset of JavaScript.
                  required: true
                source:
                  type: string
                  description: The source of the learning fact
                  example: https://www.google.com/
                  required: false
                relatedImage:
                  type: string
                  description: The related image of the learning fact
                  example: https://www.google.com/
                  required: false
                relatedLink:
                  type: string
                  description: The related link of the learning fact
                  example: https://www.google.com/
                  required: false
                creatorId:
                  type: number
                  description: The id of the creator of the learning fact
                  example: 1
                  required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningFact'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

  # LEARNING PACKAGE <> TAG API
...
  /api/package/{package_id}/tag/{tag_id}:
    delete:
      summary: Delete a tag by learning package id and tag id
      operationId: deleteTagByLearningPackageIdAndTagId
      tags:
        - Learning Packages <> Tags
      parameters:
        - name: package_id
          in: path
          description: The id of the learning package
          required: true
          schema:
            type: number
        - name: tag_id
          in: path
          description: The id of the tag
          required: true
          schema:
            type: number
      responses:
        '204':
          description: Deleted
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

  # LEARNING PACKAGE <> USER API
...
  /api/package/creator/{creator_id}:
    get:
      summary: Get all learning packages by creator id
      operationId: getLearningPackageByCreatorId
      tags:
        - Learning Packages <> Users
      parameters:
        - name: creator_id
          in: path
          description: The id of the creator
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LearningPackage'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
servers:
  [
    {
      url: http://localhost:3000,
      description: Development server
    }
  ]
components:
  schemas:
    User:
      type: object
      properties:
        userId:
          type: number
          description: The id of the user
          example: 1
        mail:
          type: string
          description: The mail of the user
          example: john.doe@gmail.com
        pseudo:
          type: string
          description: The pseudo of the user
          example: John
        password:
          type: string
          description: The password of the user
          example: 123456
        registrationDate:
          type: string
          description: The registration date of the user
          example: 2021-01-01
        profilePicture:
          type: string
          description: The profile picture of the user
          example: https://www.google.com/
        isActive:
          type: boolean
          description: The status of the user
          example: true
      required:
        - userId
        - mail
        - pseudo
        - password
        - registrationDate
        - isActive
...